Use KeyId instead of NId for TNEANet attributes
Add{Int,Str,Flt}AttrDat[NE] use the NodeH.GetKeyId(NId) and
EdgeH.GetKeyId(EId) for their index into the attribute vectors, but the
rest of the functions did not. This would allow the functions to still
work if all the nodes corresponded exactly between the NId and the
KeyId, but this doesn't happen when you use LoadEdgeList.